python_redshift_example.py
#!/usr/bin/env python3
"""
Python Redshift Connection Example
This shows how to connect to Redshift using psycopg2 with SAML authentication
"""

import psycopg2
import os
import subprocess

def get_saml_credentials():
    """Get SAML credentials for Redshift connection"""
    
    # This would typically involve calling the SAML federation endpoint
    # For now, this is a placeholder showing the concept
    
    print("🔐 SAML authentication would happen here")
    print("💡 In practice, you'd use the JDBC driver with SAML plugin")
    
    return None

def connect_to_redshift_beta():
    """Connect to Beta Redshift cluster"""
    
    # Connection parameters for Beta cluster
    conn_params = {
        'host': 'beta-wwopscomp.c0g2hsdsbjbt.us-east-1.redshift.amazonaws.com',
        'port': 5439,
        'database': 'betawwopscomp',
        'user': 'your_username',  # This would come from SAML
        'password': 'your_saml_token'  # This would come from SAML
    }
    
    try:
        # Note: This won't work without proper SAML authentication
        # This is just showing the connection pattern
        conn = psycopg2.connect(**conn_params)
        
        cursor = conn.cursor()
        cursor.execute("SELECT current_user, current_database();")
        result = cursor.fetchone()
        
        print(f"Connected as: {result[0]} to database: {result[1]}")
        
        cursor.close()
        conn.close()
        
        return True
        
    except Exception as e:
        print(f"Connection failed: {e}")
        return False

def main():
    """Main function"""
    
    print("🎯 Python Redshift Connection Example")
    print("=" * 50)
    
    print("⚠️  Note: This example shows the connection pattern")
    print("   For actual connections, use DataGrip with SAML plugin")
    print("   or AWS CLI with redshift-data commands")
    
    # Test SAML credentials
    creds = get_saml_credentials()
    
    if creds:
        connect_to_redshift_beta()
    else:
        print("❌ SAML authentication not implemented")
        print("💡 Use DataGrip with SAML plugin for actual connections")

if __name__ == "__main__":
    main()
