requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
plotly>=5.15.0
boto3>=1.26.0
python-dotenv>=1.0.0
praw>=7.7.0
numpy>=1.24.0
requests>=2.28.0